requirements.txt
cn2an==0.5.20
numpy
pyowm
jieba
colorama==0.4.6
fastasr==0.0.4
jamo==0.4.1
keyboard==0.13.5
langchain>=0.0.239
librosa
matplotlib==3.7.1
poe_api
PyAudio==0.2.13
pydantic==1.10.9
pyopenjtalk
pypinyin==0.49.0
scipy==1.11.1
soundfile==0.12.1
torch==2.0.1
Unidecode==1.3.6
faiss-cpu
google-search-results>=2.4.2
